+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## master (unreleased)
 
+### Fixes
+
+* Fix failure for `get_total_result_of_rspec_html`
+  on empty string param
+
 ## 0.1.0 (2020-03-20)
 
 ### New features