+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Fixes
 
-* Fix failure for `get_total_result_of_rspec_html`
+* Fix failure for `get_total_result_of_rspec_html`,
+  `get_failed_cases_count_from_html`  
   on empty string param
 
 ## 0.1.0 (2020-03-20)