+++ CHANGELOG.md
@@ -5,8 +5,9 @@
 ### Fixes
 
 * Fix failure for `get_total_result_of_rspec_html`,
-  `get_failed_cases_count_from_html` 
+  `get_failed_cases_count_from_html`
   on empty string param
+* Fix markdown issues
 
 ## 0.1.0 (2020-03-20)