+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master (unreleased)
 
+## 0.1.1 (2020-03-24)
+
 ### Fixes
 
 * Fix failure for `get_total_result_of_rspec_html`,